Provides a simple listing of cities within the platform.

Provides a simple listing of cities within the platform.

aq_cities(country = NULL, limit = 10000, page = NULL)

Arguments

country
Limit results by a certain country -- a two-letters code see countries() for finding code based on name.

limit
Change the number of results returned, max is 10000.

page
The page of the results to query. This can be useful if e.g. there are 2000 measurements, then first use page=1 and page=2 with limit=100 to get all measurements for your query.

Value

A results data.frame (dplyr "tbl_df") with 5 columns:

city name ("city"), country code ("country"),

number of measures in total for the city ("count"),

number of locations ("locations"),

and also an URL encoded string for the city ("cityURL") which can be useful for queries involving a city argument.

and 2 attributes, a meta data.frame (dplyr "tbl_df") with 1 line and 5 columns:

the API name ("name"),

the license of the data ("license"),

the website url ("website"),

the queried page ("page"),

the limit on the number of results ("limit"),

the number of results found on the platform for the query ("found")

and a timestamp data.frame (dplyr "tbl_df") with the query time and the last time at which the data was modified on the platform.

Details

For queries involving a city argument, the URL-encoded name of the city (as in cityURL), not its name, should be used.

Examples

not_run({
  cities <- aq_cities(country = "BA")
  cities
})