Provides a simple listing of countries within the platform.

Provides a simple listing of countries within the platform.

aq_countries(limit = 100, page = 1)

Arguments

limit

Change the number of results returned, max is 10000.

page

The page of the results to query. This can be useful if e.g. there are 2000 measurements, then first use page=1 and page=2 with limit=100 to get all measurements for your query.

Value

A results data.frame (dplyr "tbl_df") with 3 columns:

the number of measures for a country ("count"),

the ISO 3166-1 alpha-2 code of the country ("code"),

and its name ("name").

and 2 attributes, a meta data.frame (dplyr "tbl_df") with 1 line and 5 columns:

the API name ("name"),

the license of the data ("license"),

the website url ("website"),

the queried page ("page"),

the limit on the number of results ("limit"),

the number of results found on the platform for the query ("found")

and a timestamp data.frame (dplyr "tbl_df") with the query time and the last time at which the data was modified on the platform.

Details

For queries involving a country argument, the code of the country, not its name, should be used.

Examples

countries <- aq_countries()
countries

#> # A tibble: 46 × 5
#>                      name  code cities locations   count
#> *                   <chr> <chr>  <int>     <int>   <int>
#> 1               Argentina    AR      1         1    2448
#> 2                 Austria    AT    370       370  149732
#> 3               Australia    AU     56        56 1217398
#> 4  Bosnia and Herzegovina    BA     11        11  305480
#> 5              Bangladesh    BD      2         2    9658
#> 6                 Belgium    BE     98        98  200781
#> 7                 Bahrain    BH      1         1    1913
#> 8                  Brazil    BR    144       144 1886225
#> 9                  Canada    CA    157       157 1147776
#> 10                  Chile    CL    158       158 2767623
#> # ... with 36 more rows
#> 

attr(countries, "meta")

#> # A tibble: 1 × 6
#>         name   license                  website  page limit found
#>       <fctr>    <fctr>                   <fctr> <int> <int> <int>
#> 1 openaq-api CC BY 4.0 https://docs.openaq.org/     1   100    46
#> 

attr(countries, "timestamp")

#> # A tibble: 1 × 1
#>             queriedAt
#>                <dttm>
#> 1 2017-02-19 14:12:00
#>