ropenaq: Accesses Air Quality Data from the Open Data Platform OpenAQ

Allows access to air quality data from the API of the OpenAQ platform <https://docs.openaq.org/>, with the different services the API offers (getting measurements for a given query, getting latest measurements, getting lists of available countries/cities/locations).

See also