Provides a simple listing of cities within the platform.

Provides a simple listing of cities within the platform.

aq_cities(country = NULL, limit = 100, page = NULL)

Arguments

country

Limit results by a certain country -- a two-letters code see countries() for finding code based on name.

limit

Change the number of results returned, max is 10000.

page

The page of the results to query. This can be useful if e.g. there are 2000 measurements, then first use page=1 and page=2 with limit=100 to get all measurements for your query.

Value

A results data.frame (dplyr "tbl_df") with 5 columns:

city name ("city"), country code ("country"),

number of measures in total for the city ("count"),

number of locations ("locations"),

and also an URL encoded string for the city ("cityURL") which can be useful for queries involving a city argument.

and 2 attributes, a meta data.frame (dplyr "tbl_df") with 1 line and 5 columns:

the API name ("name"),

the license of the data ("license"),

the website url ("website"),

the queried page ("page"),

the limit on the number of results ("limit"),

the number of results found on the platform for the query ("found")

and a timestamp data.frame (dplyr "tbl_df") with the query time and the last time at which the data was modified on the platform.

Details

For queries involving a city argument, the URL-encoded name of the city (as in cityURL), not its name, should be used.

Examples

cities <- aq_cities(country = "IN")
cities

#> # A tibble: 36 × 5
#>          city country locations  count    cityURL
#>         <chr>   <chr>     <int>  <int>      <chr>
#> 1        Agra      IN         1  39812       Agra
#> 2   Ahmedabad      IN         1  35026  Ahmedabad
#> 3    Amritsar      IN         1   3605   Amritsar
#> 4  Aurangabad      IN         1  49392 Aurangabad
#> 5  Barddhaman      IN         3   2470 Barddhaman
#> 6   Bengaluru      IN         5 175125  Bengaluru
#> 7  Chandrapur      IN         2  99586 Chandrapur
#> 8     Chennai      IN         4 113327    Chennai
#> 9    Chittoor      IN         1   2013   Chittoor
#> 10      Delhi      IN        16 547434      Delhi
#> # ... with 26 more rows
#> 

attr(cities, "meta")

#> # A tibble: 1 × 6
#>         name   license                  website  page limit found
#>       <fctr>    <fctr>                   <fctr> <int> <int> <int>
#> 1 openaq-api CC BY 4.0 https://docs.openaq.org/     1   100    36
#> 

attr(cities, "timestamp")

#> # A tibble: 1 × 1
#>             queriedAt
#>                <dttm>
#> 1 2017-03-12 14:14:02
#>