Providing data about distinct measurement locations.

Providing data about distinct measurement locations.

aq_locations(country = NULL, city = NULL, location = NULL,
  parameter = NULL, has_geo = NULL, limit = 100, latitude = NULL,
  longitude = NULL, radius = NULL, page = NULL)

Arguments

country

Limit results by a certain country -- a two-letters code see countries() for finding code based on name.

city

Limit results by a certain city.

location

Limit results by a certain location.

parameter

Limit to only a certain parameter (valid values are 'pm25', 'pm10', 'so2', 'no2', 'o3', 'co' and 'bc'). If no parameter is given, all parameters are retrieved.

has_geo

has_geo Filter out items that have or do not have geographic information.

limit

Change the number of results returned, max is 10000.

latitude

Latitude of the center point (lat, lon) used to get measurements within a certain area.

longitude

Longitude of the center point (lat, lon) used to get measurements within a certain area.

radius

Radius (in meters) used to get measurements within a certain area, must be used with latitude and longitude

page

The page of the results to query. This can be useful if e.g. there are 2000 measurements, then first use page=1 and page=2 with limit=100 to get all measurements for your query.

Value

A results data.frame (dplyr "tbl_df") with 12 columns:

the name of the location ("location"),

the city it is in ("city"),

the code of country it is in ("country"),

the name of the source of the information ("sourceName"),

the number of measures for this location in the platform ("count"),

the last time and date at which measures were updated for this location ("lastUpdated"),

the first time and date at which measures were updated for this location ("firstUpdated"),

the parameters available for this location ("parameters"),

its longitude ("longitude") and latitude if available ("latitude"),

and finally an URL encoded version of the city name ("cityURL")

and of the location name ("locationURL").

and two attribues, a meta data.frame (dplyr "tbl_df") with 1 line and 5 columns:

the API name ("name"),

the license of the data ("license"),

the website url ("website"),

the queried page ("page"),

the limit on the number of results ("limit"),

the number of results found on the platform for the query ("found")

and a timestamp data.frame (dplyr "tbl_df") with the query time and the last time at which the data was modified on the platform.

Details

For queries involving a city or location argument, the URL-encoded name of the city/location (as in cityURL/locationURL), not its name, should be used. You can query any nested combination of country/location/city (level 1, 2 and 3), with only one value for each argument. If you write inconsistent combination such as city="Paris" and country="IN", an error message will be returned. If you write city="Delhi", you do not need to write the code of the country, unless one day there is a city with the same name in another country.

Examples

india_locations <- aq_locations(country='IN')
india_locations

#> # A tibble: 91 × 19
#>                                       location      city country  count
#>                                          <chr>     <chr>   <chr>  <int>
#> 1                        AAQMS Karve Road Pune      Pune      IN  75553
#> 2                            Alandur Bus Depot   Chennai      IN  28930
#> 3                                  Anand Vihar     Delhi      IN  61947
#> 4                                  AP Tirumala  Chittoor      IN   3172
#> 5                                Ardhali Bazar  Varanasi      IN 107235
#> 6                     Bollaram Industrial Area Hyderabad      IN     25
#> 7  Bollaram Industrial Area, Hyderabad - TSPCB Hyderabad      IN   4225
#> 8                                   BTM Layout Bengaluru      IN  54696
#> 9                        BWSSB Kadabesanahalli Bengaluru      IN  31108
#> 10                              Central School   Lucknow      IN  43857
#> # ... with 81 more rows, and 15 more variables: sourceNames <list>,
#> #   lastUpdated <dttm>, firstUpdated <dttm>, sourceName <chr>, latitude <dbl>,
#> #   longitude <dbl>, pm25 <lgl>, pm10 <lgl>, no2 <lgl>, so2 <lgl>, o3 <lgl>,
#> #   co <lgl>, bc <lgl>, cityURL <chr>, locationURL <chr>
#> 

attr(india_locations, "meta")

#> # A tibble: 1 × 6
#>         name   license                  website  page limit found
#>       <fctr>    <fctr>                   <fctr> <int> <int> <int>
#> 1 openaq-api CC BY 4.0 https://docs.openaq.org/     1   100    91
#> 

attr(india_locations, "timestamp")

#> # A tibble: 1 × 1
#>             queriedAt
#>                <dttm>
#> 1 2017-03-12 14:14:06
#> 

aq_locations(city='Houston', parameter='co')

#> # A tibble: 7 × 19
#>                                location    city country count sourceNames
#>                                   <chr>   <chr>   <chr> <int>      <list>
#> 1               Clinton C403/C304/AH113 Houston      US  2877   <chr [1]>
#> 2 Hou.DeerPrk2 C35/235/1001/AFH139FP239 Houston      US  2859   <chr [1]>
#> 3              Houston North Loop C1052 Houston      US  2873   <chr [1]>
#> 4                UH Coastal Center C697 Houston      US  2716   <chr [1]>
#> 5                   UH Moody Tower C695 Houston      US  2660   <chr [1]>
#> 6                  UH West Liberty C699 Houston      US  2796   <chr [1]>
#> 7               UH WG Jones Forest C698 Houston      US  2906   <chr [1]>
#> # ... with 14 more variables: lastUpdated <dttm>, firstUpdated <dttm>,
#> #   sourceName <chr>, latitude <dbl>, longitude <dbl>, pm25 <lgl>, pm10 <lgl>,
#> #   no2 <lgl>, so2 <lgl>, o3 <lgl>, co <lgl>, bc <lgl>, cityURL <chr>,
#> #   locationURL <chr>
#>